[MTM-18] [#92] feat: 프로젝트 소개글 등록페이지에서 생성요청하도록 변경
